# services/exchange/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: API обмена валют
  version: "1.0"
  description: API для конвертации валют и управления обменными курсами.

servers:
  - url: http://localhost:8080/api/exchange

paths:
  /convert:
    post:
      summary: "Конвертировать из одной валюты в другую"
      operationId: convertCurrency
      security:
        - bearerAuth: []
      tags:
        - Exchange
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConvertRequestDto'
      responses:
        '200':
          description: "Результат конверсии"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConvertResponseDto'

  /accept:
    post:
      summary: "Добавить/обновить курс обмена"
      operationId: updateExchangeRate
      security:
        - bearerAuth: []
      tags:
        - Exchange
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RateDto'
      responses:
        '200':
          description: "Курс сохранен"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "OK"
                    description: "Статус операции"

  /rates:
    get:
      summary: "Посмотреть все курсы обмена"
      operationId: getExchangeRates
      tags:
        - Exchange
      description: "Не требуется авторизация"
      responses:
        '200':
          description: "Список курсов"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RateDto'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ConvertRequestDto:
      type: object
      properties:
        fromCurrency:
          $ref: '#/components/schemas/Currency'
          description: "Исходная валюта"
        toCurrency:
          $ref: '#/components/schemas/Currency'
          description: "Целевая валюта"
        amount:
          type: string
          format: decimal
          example: "5000.00"
          description: "Сумма для конвертации"
      required:
        - fromCurrency
        - toCurrency
        - amount

    ConvertResponseDto:
      type: object
      properties:
        result:
          type: string
          format: decimal
          example: "5000.00"
          description: "Результат конверсии"
      required:
        - result

    RateDto:
      type: object
      properties:
        currency:
          $ref: '#/components/schemas/Currency'
          description: "Валюта"
        value:
          type: string
          format: decimal
          example: "5000.00"
          description: "Курс по отношению к базовой валюте (RUB)"
      required:
        - currency
        - value

    Currency:
      type: string
      description: "Допустимые коды валют"
      enum:
        - RUB
        - USD
        - EUR